Rework the module and function traits
