+++ crates/rune-macros/src/lib.rs
@@ -39,28 +39,6 @@ mod spanned;
 mod to_tokens;
 mod to_value;
 
-/// Macro helper function for quoting the token stream as macro output.
-///
-/// Is capable of quoting everything in Rune, except for the following:
-/// * Labels, which must be created using `Label::new`.
-/// * Dynamic quoted strings and other literals, which must be created using
-///   `Lit::new`.
-///
-/// ```
-/// use rune::macros::quote;
-///
-/// quote!(hello self);
-/// ```
-///
-/// # Interpolating values
-///
-/// Values are interpolated with `#value`, or `#(value + 1)` for expressions.
-///
-/// # Iterators
-///
-/// Anything that can be used as an iterator can be iterated over with
-/// `#(iter)*`. A token can also be used to join inbetween each iteration, like
-/// `#(iter),*`.
 #[proc_macro]
 pub fn quote(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = proc_macro2::TokenStream::from(input);
@@ -74,111 +52,6 @@ pub fn quote(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     output.into()
 }
 
-/// Macro used to annotate native functions which can be loaded into rune.
-///
-/// This macro automatically performs the following things:
-/// * Rust documentation comments are captured so that it can be used in
-///   generated Rune documentation.
-/// * The name of arguments is captured to improve documentation generation.
-/// * If an instance function is annotated this is detected (if the function
-///   receives `self`). This behavior can be forced using `#[rune(instance)]` if
-///   the function doesn't take `self`.
-///
-/// # Examples
-///
-/// A simple free function:
-///
-/// ```
-/// use rune::{Module, ContextError};
-///
-/// /// This is a pretty neat function which is called `std::str::to_uppercase("hello")`.
-/// #[rune::function]
-/// fn to_uppercase(string: &str) -> String {
-///     string.to_uppercase()
-/// }
-///
-/// fn module() -> Result<Module, ContextError> {
-///     let mut m = Module::new();
-///     m.function_meta(to_uppercase)?;
-///     Ok(m)
-/// }
-/// ```
-///
-/// A free instance function:
-///
-/// ```
-/// use rune::{Module, ContextError};
-///
-/// /// This is a pretty neat function, which is called like `"hello".to_uppercase()`.
-/// #[rune::function(instance)]
-/// fn to_uppercase(string: &str) -> String {
-///     string.to_uppercase()
-/// }
-///
-/// /// This is a pretty neat function, which is called like `string::to_uppercase2("hello")`.
-/// #[rune::function(path = string)]
-/// fn to_uppercase2(string: &str) -> String {
-///     string.to_uppercase()
-/// }
-///
-/// fn module() -> Result<Module, ContextError> {
-///     let mut m = Module::new();
-///     m.function_meta(to_uppercase)?;
-///     m.function_meta(to_uppercase2)?;
-///     Ok(m)
-/// }
-/// ```
-///
-/// A regular instance function:
-///
-/// ```
-/// use rune::{Any, Module, ContextError};
-///
-/// #[derive(Any)]
-/// struct String {
-///     inner: std::string::String
-/// }
-///
-/// impl String {
-///     /// Construct a new string wrapper.
-///     #[rune::function(path = Self::new)]
-///     fn new(string: &str) -> Self {
-///         Self {
-///             inner: string.into()
-///         }
-///     }
-///
-///     /// Construct a new string wrapper.
-///     #[rune::function(path = Self::new2)]
-///     fn new2(string: &str) -> Self {
-///         Self {
-///             inner: string.into()
-///         }
-///     }
-///
-///     /// Uppercase the string inside of the string wrapper.
-///     ///
-///     /// # Examples
-///     ///
-///     /// ```rune
-///     /// let string = String::new("hello");
-///     /// assert_eq!(string.to_uppercase(), "HELLO");
-///     /// ```
-///     #[rune::function]
-///     fn to_uppercase(&self) -> std::string::String {
-///         self.inner.to_uppercase()
-///     }
-/// }
-///
-/// fn module() -> Result<Module, ContextError> {
-///     let mut m = Module::new();
-///     m.ty::<String>()?;
-///     m.function_meta(String::new)?;
-///     m.function_meta(String::new2)?;
-///     m.function_meta(String::to_uppercase)?;
-///     Ok(m)
-/// }
-/// ```
 #[proc_macro_attribute]
 pub fn function(
     attrs: proc_macro::TokenStream,
@@ -211,7 +84,6 @@ pub fn macro_(
     output.into()
 }
 
-/// Helper derive to implement `ToTokens`.
 #[proc_macro_derive(ToTokens, attributes(rune))]
 #[doc(hidden)]
 pub fn to_tokens(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -219,7 +91,6 @@ pub fn to_tokens(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Helper derive to implement `Parse`.
 #[proc_macro_derive(Parse, attributes(rune))]
 #[doc(hidden)]
 pub fn parse(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -227,7 +98,6 @@ pub fn parse(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Helper derive to implement `Spanned`.
 #[proc_macro_derive(Spanned, attributes(rune))]
 #[doc(hidden)]
 pub fn spanned(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -235,7 +105,6 @@ pub fn spanned(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Helper derive to implement `OptionSpanned`.
 #[proc_macro_derive(OptionSpanned, attributes(rune))]
 #[doc(hidden)]
 pub fn option_spanned(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -243,7 +112,6 @@ pub fn option_spanned(input: proc_macro::TokenStream) -> proc_macro::TokenStream
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Helper derive to implement `Opaque`.
 #[proc_macro_derive(Opaque, attributes(rune))]
 #[doc(hidden)]
 pub fn opaque(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -251,37 +119,6 @@ pub fn opaque(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Derive macro for the `FromValue` trait for converting types from the dynamic
-/// `Value` container.
-///
-/// # Examples
-///
-/// ```
-/// use rune::{FromValue, Vm};
-/// use std::sync::Arc;
-///
-/// #[derive(FromValue)]
-/// struct Foo {
-///     field: u64,
-/// }
-///
-/// let mut sources = rune::sources! {
-///     entry => {
-///         pub fn main() {
-///             #{field: 42}
-///         }
-///     }
-/// };
-///
-/// let unit = rune::prepare(&mut sources).build()?;
-///
-/// let mut vm = Vm::without_runtime(Arc::new(unit));
-/// let foo = vm.call(["main"], ())?;
-/// let foo: Foo = rune::from_value(foo)?;
-///
-/// assert_eq!(foo.field, 42);
-/// # Ok::<_, rune::Error>(())
-/// ```
 #[proc_macro_derive(FromValue, attributes(rune))]
 pub fn from_value(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = syn::parse_macro_input!(input as syn::DeriveInput);
@@ -290,37 +127,6 @@ pub fn from_value(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         .into()
 }
 
-/// Derive macro for the `FromValue` trait for converting types into the dynamic
-/// `Value` container.
-///
-/// # Examples
-///
-/// ```
-/// use rune::{ToValue, Vm};
-/// use std::sync::Arc;
-///
-/// #[derive(ToValue)]
-/// struct Foo {
-///     field: u64,
-/// }
-///
-/// let mut sources = rune::sources! {
-///     entry => {
-///         pub fn main(foo) {
-///             foo.field + 1
-///         }
-///     }
-/// };
-///
-/// let unit = rune::prepare(&mut sources).build()?;
-///
-/// let mut vm = Vm::without_runtime(Arc::new(unit));
-/// let value = vm.call(["main"], (Foo { field: 42 },))?;
-/// let value: u64 = rune::from_value(value)?;
-///
-/// assert_eq!(value, 43);
-/// # Ok::<_, rune::Error>(())
-/// ```
 #[proc_macro_derive(ToValue, attributes(rune))]
 pub fn to_value(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = syn::parse_macro_input!(input as syn::DeriveInput);
@@ -329,40 +135,12 @@ pub fn to_value(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         .into()
 }
 
-/// Macro to mark a value as external, which will implement all the appropriate
-/// traits.
-///
-/// This is required to support the external type as a type argument in a
-/// registered function.
-///
-/// ## `#[rune(name = "..")]` attribute
-///
-/// The name of a type defaults to its identifiers, so `struct Foo {}` would be
-/// given the name `"Foo"`.
-///
-/// This can be overrided with the `#[rune(name = "...")]` attribute:
-///
-/// ```
-/// use rune::Any;
-///
-/// #[derive(Any)]
-/// #[rune(name = "Bar")]
-/// struct Foo {
-/// }
-///
-/// fn install() -> Result<rune::Module, rune::ContextError> {
-///     let mut module = rune::Module::new();
-///     module.ty::<Foo>()?;
-///     Ok(module)
-/// }
-/// ```
 #[proc_macro_derive(Any, attributes(rune))]
 pub fn any(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let derive = syn::parse_macro_input!(input as any::Derive);
     derive.expand().unwrap_or_else(to_compile_errors).into()
 }
 
-/// Internal macro to implement external.
 #[proc_macro]
 #[doc(hidden)]
 pub fn __internal_impl_any(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -373,7 +151,6 @@ pub fn __internal_impl_any(input: proc_macro::TokenStream) -> proc_macro::TokenS
         .into()
 }
 
-/// Internal macro to instrument a function which is threading AST.
 #[proc_macro_attribute]
 #[doc(hidden)]
 pub fn __instrument_ast(
