+++ crates/rune/src/lib.rs
@@ -224,8 +224,6 @@ pub use self::compile::{Context, ContextError, Options};
 pub mod module;
 #[doc(inline)]
 pub use self::module::module::Module;
-#[doc(inline)]
-pub use self::module::InstallWith;
 
 pub mod diagnostics;
 #[doc(inline)]
@@ -269,16 +267,132 @@ cfg_workspace! {
 
 // Macros used internally and re-exported.
 pub(crate) use rune_macros::__internal_impl_any;
-pub use rune_macros::{function, macro_};
 
-#[cfg(feature = "cli")]
-pub mod cli;
+/// Macro used to annotate native functions which can be loaded into rune.
+///
+/// This macro automatically performs the following things:
+/// * Rust documentation comments are captured so that it can be used in
+///   generated Rune documentation.
+/// * The name of arguments is captured to improve documentation generation.
+/// * If an instance function is annotated this is detected (if the function
+///   receives `self`). This behavior can be forced using `#[rune(instance)]` if
+///   the function doesn't take `self`.
+///
+/// # Examples
+///
+/// A simple free function:
+///
+/// ```
+/// use rune::{Module, ContextError};
+///
+/// /// This is a pretty neat function which is called `std::str::to_uppercase("hello")`.
+/// #[rune::function]
+/// fn to_uppercase(string: &str) -> String {
+///     string.to_uppercase()
+/// }
+///
+/// fn module() -> Result<Module, ContextError> {
+///     let mut m = Module::new();
+///     m.function_meta(to_uppercase)?;
+///     Ok(m)
+/// }
+/// ```
+///
+/// A free instance function:
+///
+/// ```
+/// use rune::{Module, ContextError};
+///
+/// /// This is a pretty neat function, which is called like `"hello".to_uppercase()`.
+/// #[rune::function(instance)]
+/// fn to_uppercase(string: &str) -> String {
+///     string.to_uppercase()
+/// }
+///
+/// /// This is a pretty neat function, which is called like `string::to_uppercase2("hello")`.
+/// #[rune::function(path = string)]
+/// fn to_uppercase2(string: &str) -> String {
+///     string.to_uppercase()
+/// }
+///
+/// fn module() -> Result<Module, ContextError> {
+///     let mut m = Module::new();
+///     m.function_meta(to_uppercase)?;
+///     m.function_meta(to_uppercase2)?;
+///     Ok(m)
+/// }
+/// ```
+///
+/// A regular instance function:
+///
+/// ```
+/// use rune::{Any, Module, ContextError};
+///
+/// #[derive(Any)]
+/// struct String {
+///     inner: std::string::String
+/// }
+///
+/// impl String {
+///     /// Construct a new string wrapper.
+///     #[rune::function(path = Self::new)]
+///     fn new(string: &str) -> Self {
+///         Self {
+///             inner: string.into()
+///         }
+///     }
+///
+///     /// Construct a new string wrapper.
+///     #[rune::function(path = Self::new2)]
+///     fn new2(string: &str) -> Self {
+///         Self {
+///             inner: string.into()
+///         }
+///     }
+///
+///     /// Uppercase the string inside of the string wrapper.
+///     ///
+///     /// # Examples
+///     ///
+///     /// ```rune
+///     /// let string = String::new("hello");
+///     /// assert_eq!(string.to_uppercase(), "HELLO");
+///     /// ```
+///     #[rune::function]
+///     fn to_uppercase(&self) -> std::string::String {
+///         self.inner.to_uppercase()
+///     }
+/// }
+///
+/// fn module() -> Result<Module, ContextError> {
+///     let mut m = Module::new();
+///     m.ty::<String>()?;
+///     m.function_meta(String::new)?;
+///     m.function_meta(String::new2)?;
+///     m.function_meta(String::to_uppercase)?;
+///     Ok(m)
+/// }
+/// ```
+#[doc(hidden)]
+pub use rune_macros::function;
+
+/// Macro used to annotate native functions which can be loaded as macros in
+/// rune.
+///
+/// See [`Module::macro_meta`].
+#[doc(hidden)]
+pub use rune_macros::macro_;
+
+cfg_cli! {
+    pub mod cli;
+}
 
 #[cfg(feature = "languageserver")]
 pub mod languageserver;
 
-#[cfg(feature = "doc")]
-pub mod doc;
+cfg_doc! {
+    pub mod doc;
+}
 
 /// Internal collection re-export.
 mod collections {
