+++ crates/rune/src/parse/opaque.rs
@@ -1,5 +1,7 @@
 use crate::ast::Spanned;
 use crate::parse::{Id, NonZeroId};
+
+/// Helper derive to implement [`Opaque`].
 pub(crate) use rune_macros::Opaque;
 
 pub(crate) trait Opaque {
@@ -8,12 +10,14 @@ pub(crate) trait Opaque {
 }
 
 impl Opaque for Id {
+    #[inline]
     fn id(&self) -> Id {
         *self
     }
 }
 
 impl Opaque for NonZeroId {
+    #[inline]
     fn id(&self) -> Id {
         Id::new(*self)
     }
@@ -23,6 +27,7 @@ impl<T> Opaque for &T
 where
     T: Opaque,
 {
+    #[inline]
     fn id(&self) -> Id {
         (*self).id()
     }
@@ -33,6 +38,7 @@ where
     S: Spanned,
     O: Opaque,
 {
+    #[inline]
     fn id(&self) -> Id {
         self.1.id()
     }
